+++ cuda/src/include/utils.h
@@ -24,9 +24,9 @@ void load_parameters_from_file(char filename[], int *N_PART, int *BINS, double *
                                unsigned int *Ntandas, int steps[], char inputFilename[], char saveFilename[],
                                bool *resume, bool *dump, double *sigmaL);
 
-void read_data(char filename[], double *x, double *p, unsigned int *evolution, int N_PART);
+void read_data(char filename[], float *x, double *p, unsigned int *evolution, int N_PART);
 
-void save_data(char filename[], double *x, double *p, unsigned int evolution, int N_PART);
+void save_data(char filename[], float *x, double *p, unsigned int evolution, int N_PART);
 
 double energy_sum(double *p, int N_PART, unsigned int evolution, double M);
